+++ app/Events/Bootstrap.php
@@ -3,6 +3,7 @@
 namespace App\Events;
 
 use Bayfront\Bones\Abstracts\EventSubscriber;
+use Bayfront\Bones\Application\Services\Events\EventSubscription;
 use Bayfront\Bones\Application\Utilities\App;
 use Bayfront\Bones\Interfaces\EventSubscriberInterface;
 use Bayfront\HttpResponse\Response;
@@ -31,12 +32,7 @@ class Bootstrap extends EventSubscriber implements EventSubscriberInterface
     public function getSubscriptions(): array
     {
         return [
-            'app.bootstrap' => [
-                [
-                    'method' => 'modifyResponseHeaders',
-                    'priority' => 5
-                ]
-            ],
+            new EventSubscription('app.bootstrap', [$this, 'modifyResponseHeaders'], 10)
         ];
     }