+++ config/app.php
@@ -8,19 +8,6 @@ return [
     'debug' => App::getEnv('APP_DEBUG'),
     'environment' => App::getEnv('APP_ENVIRONMENT'), // e.g.: "dev", "staging", "qa", "prod"
     'timezone' => App::getEnv('APP_TIMEZONE'), // See: https://www.php.net/manual/en/timezones.php
-    'events' => [
-        'autoload' => true,
-        'load' => []
-    ],
-    'filters' => [
-        'autoload' => true,
-        'load' => []
-    ],
-    'commands' => [
-        'autoload' => true,
-        'load' => []
-    ],
-    'backup_path' => App::getEnv('APP_BACKUP_PATH'), // Optional. Used with deploy: commands
     'locale' => [ // Locale settings
         'valid' => [
             'en',
@@ -42,5 +29,5 @@ return [
             ]
         ]
     ],
-    'version' => '1.0.0' // App version
+    'version' => '1.0.0' // Current app version (string)
 ];