+++ CHANGELOG.md
@@ -12,6 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities
 
+## [2.0.0] - 2024.02.03
+
+### Changed
+
+- Updated to work with Bones `v4.0`
+- Moved deploy config to `config/deploy.php`
+
 ## [1.0.2] - 2023.08.10
 
 ### Changed